+++ quickstart/src/main/kotlin/jp/shiguredo/sora/quickstart/MainActivity.kt
@@ -4,9 +4,9 @@ import android.Manifest
 import android.graphics.Color
 import android.media.AudioManager
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AlertDialog
-import android.support.v7.app.AppCompatActivity
+import com.google.android.material.snackbar.Snackbar
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import jp.shiguredo.sora.sdk.camera.CameraCapturerFactory
 import jp.shiguredo.sora.sdk.channel.SoraMediaChannel